refactor: move reasoning_helpers into base/helpers/; rename orphan_fix (#310)
- Move converters/reasoning_helpers.py → converters/base/helpers/reasoning.py
- Rename orphan_fix.py → tool_orphan_fix.py for consistent tool_* prefix

helpers/ now organized by target type:
- tool: tool_orphan_fix, tool_call_unwind, tool_content
- image: image_limit
- reasoning: reasoning
- general: cache, schema
