+++ src/llm_rosetta/converters/base/helpers/tool_orphan_fix.py
